requirements.txt
openai>=1.80.0
faiss-cpu>=1.11.0
python-dotenv>=1.1.0
PyPDF2>=3.0.1
numpy>=2.0.0
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.45.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
protobuf>=3.20,<7
